Db connection Guard, connection pool, create problem route

+++ src/models/problem.rs
@@ -1,11 +1,13 @@
 use chrono::{DateTime, Utc};
+use serde::Serialize;
 
-#[derive(Queryable)]
+#[derive(Queryable, Serialize)]
+#[serde(rename_all = "camelCase")]
 pub struct Problem {
     pub id: i32,
     pub title: String,
-    pub grade: u8,
-    pub rating: u8,
+    pub grade: i32,
+    pub rating: i32,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
